Remove filterArgs()
I previously added a 'hacky' filterArgs() to remove the unnecessary
parameters from the GET request by setting the 'name' parameter
to null for empty args, but it's broken now and I need to
remove the empty params and reconstruct the URL manually using
javascript. I'd rather just let them be for now.